requirements.txt
langchain_openai
langchain_core
streamlit
python-dotenv
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
langchainhub
groq
cassio
beautifulsoup4
langchain-groq
langchain-objectbox
unstructured
ipykernel
sentence_transformers
langchain-openai
requests